index 수정 0213 오전본

+++ food_hub_project/src/main/java/com/application/foodhub/post/PostService.java
@@ -36,6 +36,6 @@ public interface PostService {
 
 	public String getCategoryName(Long categoryId);
 	
-	public List<Map<String, Object>> getLatestPostsByCategoryId(int categoryId);
+	public List<Map<String, Object>> getLatestPostsByCategoryId(long categoryId);
 
 }
